moved static helper methods out of PatternMatcher/NonACPatternMatcher
