+++ java-backend/src/main/java/org/kframework/backend/java/util/Utils.java
@@ -1,8 +1,6 @@
 // Copyright (c) 2013-2014 K Team. All Rights Reserved.
 package org.kframework.backend.java.util;
 
-
-
 /**
  * Class grouping various useful constants and static methods.
  *
